requirements.txt
paho-mqtt==2.1.0
requests==2.32.3
flask==3.0.3
db-sqlite3==0.0.1
pandas==2.2.2
flask_cors==4.0.1
